[level 2] Title: 택배 배달과 수거하기, Time: 45.69 ms, Memory: 73.8 MB -BaekjoonHub

+++ 프로그래머스/2/150369. 택배 배달과 수거하기/택배 배달과 수거하기.java
@@ -22,7 +22,7 @@ class Solution {
             int farPickup = pickupStack.isEmpty() ? -1 : pickupStack.peek();
 
             int farHouse = Math.max(farDelivery, farPickup);
-            answer += (long) (farHouse + 1) * 2;
+            answer += (farHouse + 1) * 2;
 
             int deliveryBox = 0;
             while (!deliveryStack.isEmpty() && deliveryBox < cap) {
